+++ StudentEnrolment/StudentEnrolment/Controllers/SubjectController.cs
@@ -1,6 +1,11 @@
 ﻿using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 using StudentEnrolment.Data;
 using StudentEnrolment.Models;
+using StudentEnrolment.Models.BaseClasses;
+using StudentEnrolment.Models.DTOModels;
+using StudentEnrolment.Models.EntityModels;
+using System.Linq;
 
 namespace StudentEnrolment.Controllers
 {
@@ -15,43 +20,80 @@ namespace StudentEnrolment.Controllers
         [HttpGet("Subject/List")]
         public List<Subject> List()
         {
-            List<Subject> itemList = _db.Subject.ToList();
+            List<Subject> itemList = _db.Subject
+                                    .Include(c => c.CourseSubject)
+                                    .ThenInclude(cs => cs.Course)
+                                    .ToList();
             return itemList;
         }
 
         [HttpPost("Subject/Create")]
-        public IActionResult Create([FromBody] Subject subject)
+        public IActionResult Create([FromBody] SubjectDto subjectDto)
         {
+            if (subjectDto == null)
+                return BadRequest();
+
+            Guid guid = Guid.NewGuid();
+
+            Subject subject = new Subject(
+                guid.ToString(),
+                subjectDto.Name,
+                subjectDto.Description
+            );
+
             try
             {
                 _db.Subject.Add(subject);
+
+                for (var i = 0; i < subjectDto.Courses.Count; i++)
+                {
+                    subjectDto.Courses[i].SubjectId = guid.ToString();
+
+                    CourseSubject courseSubject =
+                        new CourseSubject(
+                            subjectDto.Courses[i].CourseId,
+                            subjectDto.Courses[i].SubjectId
+                        );
+
+                    _db.CourseSubject.Add(courseSubject);
+                }
+
                 _db.SaveChanges();
                 return Ok();
             }
             catch (Exception ex)
             {
-                return BadRequest(ex);
+               return BadRequest(ex.Message);
             }
         }
 
         [HttpPost("Subject/Delete")]
-        public IActionResult Delete([FromBody] IdBase subject)
+        public IActionResult Delete([FromBody] List<IdBase> subjects)
         {
-            Subject foundSubject = _db.Subject.FirstOrDefault(x => x.Id == subject.Id);
-            if (foundSubject != null)
+            if (subjects == null)
+                return BadRequest("subjects is null");
+
+            for (int i = 0; i < subjects.Count; i++)
             {
-                try
+                Subject foundSubject = _db.Subject.FirstOrDefault(s => s.Id == subjects[i].Id);
+                if (foundSubject != null)
                 {
-                    _db.Subject.Remove(foundSubject);
-                    _db.SaveChanges();
-                    return Ok();
+                    try
+                    {
+                        _db.Subject.Remove(foundSubject);
+                    }
+                    catch (Exception ex)
+                    {
+                       return BadRequest(ex.Message);
+                    }
                 }
-                catch (Exception ex)
+                else
                 {
-                    return BadRequest(ex);
+                    return BadRequest("Id not found");
                 }
             }
-            return BadRequest();
+            _db.SaveChanges();
+            return Ok();
         }
     }
 }
